requirements.txt
Django>=4.2,<5
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
python-dotenv>=1.0,<2
